fix bug dawn pytorch
